+++ xonsh/cli_utils.py
@@ -12,6 +12,9 @@ import sys
 import typing as tp
 from collections import defaultdict
 
+from xonsh.built_ins import XSH
+from xonsh.completers.tools import RichCompletion
+
 TYPING_ANNOTATED_AVAILABLE = False
 """One can import ``Annotated`` from this module
 which adds a stub when it is not available in ``typing``/``typing_extensions`` modules."""
@@ -279,7 +282,6 @@ class RstHelpFormatter(ap.RawTextHelpFormatter):
 
 
 def get_argparse_formatter_class():
-    from xonsh.built_ins import XSH
     from xonsh.platform import HAS_PYGMENTS
 
     if (
@@ -384,6 +386,145 @@ def dispatch(parser: ap.ArgumentParser, args=None, lenient=False, **ns):
     return _dispatch_func(func, ns)
 
 
+class ArgparseCompleter:
+    """A completer function for ArgParserAlias commands"""
+
+    def __init__(self, parser: ap.ArgumentParser, command, **kwargs):
+        args = tuple(c.value for c in command.args[: command.arg_index])
+
+        self.parser, self.remaining_args = self.get_parser(parser, args[1:])
+
+        self.command = command
+        kwargs["command"] = command
+        self.kwargs = kwargs
+        """will be sent to completer function"""
+
+    @staticmethod
+    def get_parser(parser, args) -> tp.Tuple[ap.ArgumentParser, tp.Tuple[str, ...]]:
+        """Check for sub-parsers"""
+        sub_parsers = {}
+        for act in parser._get_positional_actions():
+            if act.nargs == ap.PARSER:
+                sub_parsers = act.choices  # there should be only one subparser
+        if sub_parsers:
+            for idx, pos in enumerate(args):
+                if pos in sub_parsers:
+                    # get the correct parser
+                    return ArgparseCompleter.get_parser(
+                        sub_parsers[pos], args[idx + 1 :]
+                    )
+        # base scenario
+        return parser, args
+
+    def filled(self, act: ap.Action) -> int:
+        """Consume remaining_args for the given action"""
+        args_len = 0
+        for arg in self.remaining_args:
+            if arg and arg[0] in self.parser.prefix_chars:
+                # stop when other --option explicitly given
+                break
+            args_len += 1
+        nargs = (
+            act.nargs
+            if isinstance(act.nargs, int)
+            else args_len + 1
+            if act.nargs in {ap.ONE_OR_MORE, ap.ZERO_OR_MORE}
+            else 1
+        )
+        if len(self.remaining_args) >= nargs:
+            # consume n-number of args
+            self.remaining_args = self.remaining_args[nargs:]
+            # complete for next action
+            return True
+        return False
+
+    def _complete(self, act: ap.Action, **kwargs):
+        if hasattr(act, "completer") and callable(act.completer):  # type: ignore
+            # call the completer function
+            kwargs.update(self.kwargs)
+            yield from act.completer(xsh=XSH, action=act, completer=self, **kwargs)  # type: ignore
+
+        if getattr(act, "choices", None) and not isinstance(act.choices, dict):
+            # any sequence or iterable
+            yield from act.choices
+
+    def _complete_pos(self, act):
+        # even subparserAction can have completer attribute set
+        yield from self._complete(act)
+
+        if isinstance(act.choices, dict):  # sub-parsers
+            for choice, sub_parser in act.choices.items():
+                yield RichCompletion(
+                    choice,
+                    description=sub_parser.description or "",
+                    append_space=True,
+                )
+
+    def complete(self):
+        # options will come before/after positionals
+        options = {act: None for act in self.parser._get_optional_actions()}
+
+        # remove options that are already filled
+        opt_completions = self._complete_options(options)
+        if opt_completions:
+            yield from opt_completions
+            return
+
+        for act in self.parser._get_positional_actions():
+            # number of arguments it consumes
+            if self.filled(act):
+                continue
+            yield from self._complete_pos(act)
+            # close after a valid positional arg completion
+            break
+
+        opt_completions = self._complete_options(options)
+        if opt_completions:
+            yield from opt_completions
+            return
+
+        # complete remaining options only if requested or enabled
+        show_opts = XSH.env.get("ALIAS_COMPLETIONS_OPTIONS_BY_DEFAULT", False)
+        if not show_opts:
+            if not (
+                self.command.prefix
+                and self.command.prefix[0] in self.parser.prefix_chars
+            ):
+                return
+
+        # in the end after positionals show remaining unfilled options
+        for act in options:
+            for flag in act.option_strings:
+                desc = ""
+                if act.help:
+                    formatter = self.parser._get_formatter()
+                    try:
+                        desc = formatter._expand_help(act)
+                    except KeyError:
+                        desc = act.help
+                yield RichCompletion(flag, description=desc)
+
+    def _complete_options(self, options):
+        while self.remaining_args:
+            arg = self.remaining_args[0]
+            act_res = self.parser._parse_optional(arg)
+            if not act_res:
+                # it is not a option string: pass
+                break
+            # it is a valid option and advance
+            self.remaining_args = self.remaining_args[1:]
+            act, _, value = act_res
+
+            # remove the found option
+            # todo: not remove if append/extend
+            options.pop(act, None)
+
+            if self.filled(act):
+                continue
+            # stop suggestion until current option is complete
+            return self._complete(act)
+
+
 class ArgParserAlias:
     """Provides a structure to the Alias. The parser is lazily loaded.
 
@@ -426,6 +567,10 @@ class ArgParserAlias:
             add_args(parser, func, allowed_params=allowed_params)
         return parser
 
+    def xonsh_complete(self, command, **kwargs):
+        completer = ArgparseCompleter(self.parser, command=command, **kwargs)
+        yield from completer.complete()
+
     def __call__(
         self,
         args,
@@ -453,6 +598,7 @@ class ArgParserAlias:
 __all__ = (
     "Arg",
     "ArgParserAlias",
+    "ArgparseCompleter",
     "Annotated",
     "ArgParser",
     "make_parser",
